// src/app/model/supply.ts
export class Supply {
    id: number;
    user_id: number;
    wet_product_id: number;
    dry_product_id: number;
    supplier_code: string;
    product: string;
    inventory_received: number;
    transport_cost: number;
    purchase_price: number;
    purchase_date: Date;
    supply_price: number;
    supply_date: Date;
}